# pkg/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: CMDB API
  version: 0.1.0
  description: 纯自研 CMDB 核心 API 契约，契约先行，前后端据此并行开发
servers:
  - url: /
    description: 相对路径（与后端同域部署）
...
security:
  - cookieAuth: []
  - bearerAuth: []
paths:
  /api/v1/auth/login:
    post:
      tags: [auth]
      summary: 用户名密码登录
      operationId: login
      security: []
      description: 校验通过后签发 JWT,写入 httpOnly cookie(meridian_token),同时返回当前用户信息;客户端也可从响应体取 token 走 Bearer 方式
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/auth/logout:
    post:
      tags: [auth]
      summary: 退出登录
      operationId: logout
      description: 清除会话 cookie(JWT 无状态,服务端不做吊销)
      responses:
        '200':
          description: 已退出
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status:
                    type: string
                    example: ok
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/auth/me:
    get:
      tags: [auth]
      summary: 查询当前登录用户
      operationId: getCurrentUser
      responses:
        '200':
          description: 当前用户信息(含角色与权限点并集)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrentUser'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/users:
    get:
      tags: [users]
      summary: 查询用户列表
      operationId: listUsers
      parameters:
        - name: keyword
          in: query
          description: 按用户名或显示名模糊过滤
          required: false
          schema:
            type: string
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: 用户分页列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [users]
      summary: 创建用户
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreateRequest'
      responses:
        '200':
          description: 创建成功,返回完整用户
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/users/{user_id}:
    get:
      tags: [users]
      summary: 查询用户详情
      operationId: getUser
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: 用户详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [users]
      summary: 更新用户(显示名/状态/重置密码/角色分配)
      operationId: patchUser
      description: 内置 admin 用户(is_builtin=true)不允许停用、不允许移除 admin 角色
      parameters:
        - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserPatchRequest'
      responses:
        '200':
          description: 更新成功,返回完整用户
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/roles:
    get:
      tags: [roles]
      summary: 查询角色列表
      operationId: listRoles
      responses:
        '200':
          description: 角色列表(含权限点与用户数)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [roles]
      summary: 创建角色
      operationId: createRole
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleCreateRequest'
      responses:
        '200':
          description: 创建成功,返回完整角色
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Role'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/roles/{role_id}:
    get:
      tags: [roles]
      summary: 查询角色详情
      operationId: getRole
      parameters:
        - $ref: '#/components/parameters/RoleId'
      responses:
        '200':
          description: 角色详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Role'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [roles]
      summary: 更新角色(名称/描述/权限点)
      operationId: patchRole
      description: 内置角色(is_builtin=true)的 code 不可修改;内置 admin 角色的权限点不可修改
      parameters:
        - $ref: '#/components/parameters/RoleId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RolePatchRequest'
      responses:
        '200':
          description: 更新成功,返回完整角色
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Role'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [roles]
      summary: 删除角色
      operationId: deleteRole
      description: 内置角色不可删除;仍有用户关联的角色不可删除
      parameters:
        - $ref: '#/components/parameters/RoleId'
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status:
                    type: string
                    example: ok
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/permissions:
    get:
      tags: [roles]
      summary: 查询权限点目录
      operationId: listPermissions
      description: 返回系统全部权限点(固定目录,角色编辑界面据此渲染勾选项)
      responses:
        '200':
          description: 权限点列表
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/PermissionItem'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/models:
    get:
      tags: [models]
      summary: 查询模型列表
      operationId: listModels
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: keyword
          in: query
          description: 按名称或编码模糊过滤
          required: false
          schema:
            type: string
      responses:
        '200':
          description: 模型分页列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
    post:
      tags: [models]
      summary: 创建模型
      operationId: createModel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModelCreateRequest'
      responses:
        '200':
          description: 创建成功，返回完整模型
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Model'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/v1/models/{model_id}:
    get:
      tags: [models]
      summary: 查询模型详情
      operationId: getModel
      parameters:
        - $ref: '#/components/parameters/ModelId'
      responses:
        '200':
          description: 模型详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Model'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [models]
      summary: 更新模型（部分字段）
      operationId: patchModel
      parameters:
        - $ref: '#/components/parameters/ModelId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModelPatchRequest'
      responses:
        '200':
          description: 更新成功，返回完整模型
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Model'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/cis:
    get:
      tags: [cis]
      summary: 查询 CI 列表（支持模型过滤、关键字搜索与分页）
      operationId: listCIs
      parameters:
        - name: model_id
          in: query
          description: 按所属模型过滤
          required: false
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          description: 按状态过滤
          required: false
          schema:
            type: string
            enum: [discovered, active, retired]
        - name: keyword
          in: query
          description: 全文关键字（匹配全部属性值，大小写不敏感）
          required: false
          schema:
            type: string
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: CI 分页列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CIListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
    post:
      tags: [cis]
      summary: 创建 CI
      operationId: createCI
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CICreateRequest'
      responses:
        '200':
          description: 创建成功，返回完整 CI
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CI'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/v1/cis/{ci_id}:
    get:
      tags: [cis]
      summary: 查询 CI 详情
      operationId: getCI
      parameters:
        - $ref: '#/components/parameters/CIId'
      responses:
        '200':
          description: CI 详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CI'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [cis]
      summary: 更新 CI（部分字段）
      operationId: patchCI
      parameters:
        - $ref: '#/components/parameters/CIId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CIPatchRequest'
      responses:
        '200':
          description: 更新成功，返回完整 CI
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CI'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/cis/{ci_id}/relations:
    get:
      tags: [cis]
      summary: 查询 CI 的关系列表
      operationId: listCIRelations
      parameters:
        - $ref: '#/components/parameters/CIId'
      responses:
        '200':
          description: 关系列表（含对端 CI 摘要）
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/CIRelation'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags: [cis]
      summary: 创建 CI 关系
      operationId: createCIRelation
      description: 在当前 CI 与对端 CI 间建立关系；关系编码须在所属模型的关系定义内，对端 CI 模型须匹配 target_model。cardinality 为 one_to_one 时自动替换同编码同方向的既有关系（用于改挂，如机柜换机房）
      parameters:
        - $ref: '#/components/parameters/CIId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CIRelationCreateRequest'
      responses:
        '200':
          description: 创建成功，返回关系（含对端 CI 摘要）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CIRelation'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/cis/{ci_id}/relations/{relation_code}/{peer_ci_id}:
    delete:
      tags: [cis]
      summary: 删除 CI 关系
      operationId: deleteCIRelation
      parameters:
        - $ref: '#/components/parameters/CIId'
        - name: relation_code
          in: path
          description: 关系编码
          required: true
          schema:
            type: string
        - name: peer_ci_id
          in: path
          description: 对端 CI ID
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status:
                    type: string
                    example: ok
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/discovery-records:
    post:
      tags: [discovery]
      summary: 批量写入发现记录
      operationId: createDiscoveryRecords
      description: 采集器上报标准发现记录，服务端接收后写入 JetStream 等待调和
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DiscoveryRecordBatchRequest'
      responses:
        '200':
          description: 批量接收结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DiscoveryRecordBatchResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/v1/reconcile/preview:
    post:
      tags: [reconcile]
      summary: 调和预览（不实际落库）
      operationId: previewReconcile
      description: 对一条发现记录执行调和规则演练，返回将发生的动作与依据，便于规则调试
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReconcilePreviewRequest'
      responses:
        '200':
          description: 调和预览结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReconcilePreviewResult'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/v1/discovery-pool:
    get:
      tags: [discovery-pool]
      summary: 查询发现池列表
      operationId: listDiscoveryPool
      description: 调和判定为冲突或暂无法调和的发现记录在此等待人工裁决，列表关联原始记录与调和结果
      parameters:
        - name: status
          in: query
          description: 按裁决状态过滤
          required: false
          schema:
            type: string
            enum: [pending, confirmed, ignored]
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: 发现池分页列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DiscoveryPoolListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/v1/discovery-pool/{item_id}/confirm:
    post:
      tags: [discovery-pool]
      summary: 确认池条目并建档 CI
      operationId: confirmDiscoveryPoolItem
      description: 以池条目的发现属性（可被 body 覆盖）创建 status=active 的 CI，并把池条目置 confirmed
      parameters:
        - $ref: '#/components/parameters/PoolItemId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DiscoveryPoolConfirmRequest'
      responses:
        '201':
          description: 建档成功，返回完整 CI（status=active）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CI'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/discovery-pool/{item_id}/ignore:
    post:
      tags: [discovery-pool]
      summary: 忽略池条目
      operationId: ignoreDiscoveryPoolItem
      description: 把池条目置 ignored（不建档），仅 pending 状态可操作
      parameters:
        - $ref: '#/components/parameters/PoolItemId'
      responses:
        '200':
          description: 忽略成功，返回更新后的池条目
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DiscoveryPoolItem'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/search:
    get:
      tags: [cis]
      summary: 全局全文搜索
      operationId: globalSearch
      description: 跨模型、CI、IPAM 前缀/IP 的统一搜索入口（landing 页全局搜索框）。结果按资源类型分组，每组上限由 limit 控制；无对应权限点的分组自动省略（如用户无 ipam:read 则不返回 IPAM 组）
      parameters:
        - name: q
          in: query
          description: 搜索关键字（大小写不敏感，子串匹配）
          required: true
          schema:
            type: string
            minLength: 1
        - name: limit
          in: query
          description: 每组返回上限，默认 10、最大 50
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 10
      responses:
        '200':
          description: 分组搜索结果（仅返回非空分组）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/dcim/overview:
    get:
      tags: [dcim]
      summary: DCIM 容量总览
      operationId: getDCIMOverview
      description: 按机房聚合机柜数量、U 位容量/占用与电力容量，并给出全局合计；未分配机房的机柜单独归入 unassigned
      responses:
        '200':
          description: 容量总览
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DCIMOverview'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/credentials:
    get:
      tags: [credentials]
      summary: 查询凭据列表
      operationId: listCredentials
      description: 凭据元数据分页列表（secret 密文存储，任何响应永不回明文）
      parameters:
        - name: type
          in: query
          description: 按凭据类型过滤
          required: false
          schema:
            type: string
            enum: [vcenter, aliyun, volc, snmp, db, kubeconfig, ssh_ipmi, n9e, netbox]
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: 凭据分页列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CredentialListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [credentials]
      summary: 创建凭据
      operationId: createCredential
      description: secret 明文仅在请求体传入，服务端以 AES-256-GCM 加密后落库并记 create 审计
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CredentialCreateRequest'
      responses:
        '201':
          description: 创建成功，返回凭据元数据（不含 secret）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Credential'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/credentials/{credential_id}:
    patch:
      tags: [credentials]
      summary: 修改凭据非密字段
      operationId: patchCredential
      description: 仅可修改 name/description；修改 secret 请走轮换接口
      parameters:
        - $ref: '#/components/parameters/CredentialId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CredentialPatchRequest'
      responses:
        '200':
          description: 修改成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Credential'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: meridian_token
      description: 登录成功后由服务端写入的 httpOnly JWT 会话 cookie
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 采集器等非浏览器客户端使用的 Bearer JWT(登录接口响应体中的 token)
  parameters:
...
    Page:
      name: page
      in: query
      description: 页码，从 1 开始
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: page_size
      in: query
      description: 每页条数
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 20
...
  responses:
    BadRequest:
      description: 请求参数或校验规则不通过
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: 未登录或会话已失效(错误码 UNAUTHORIZED)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          example: admin
        password:
          type: string
          format: password
    LoginResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
          description: JWT 会话令牌,与 httpOnly cookie 同值,供 Bearer 方式使用
        user:
          $ref: '#/components/schemas/CurrentUser'
...